+++ src/app/SFXBasicEditor.hpp
@@ -34,7 +34,7 @@ class BasicEditorWorkSpace : public ToolWorkSpace
 public:
     void init(App* app) override;
     void onEvent(const KeyEvent& event) override;
-    void update(double t) override;
+    void update(float t) override;
     void render() override;
 
 private:
